Bugfix for save-method
A refactoring in in the previous version introduced a bug into class /USI/CL_BAL_LSTATE_ACTIVE: The attributes highest_message_number and messages were merged into the attribute messages.

The attribute messages was cleared after saving. As a result, the currently highest message number was lost and the count started from 1 again.

This then created duplicate keys on the database when the save-method was called the second time, as the message number is part of the primary key.

As the insert was done via "INSERT [...] FROM TABLE" this error resulted in a short dump.