Catch up with "internal version" (See details)
----------------------------- Bugfix -----------------------------
A bug was discovered in method /USI/CL_BAL_LOGGER->/USI/IF_BAL_LOGGER~ADD_FREE_TEXT: Passing a variable having e.g. type CHAR50 caused a short dump (runtime error CALL_METHOD_CONFLICT_TYPE).

The problem was, that the logger is delegating the call to the state object. The logger declares I_FREE_TEXT as a CSEQUENCE, the state uses type CHAR200. ABAP does obviously _NOT_ convert that at runtime. Passing a literal, that is shorter than 200 characters, worked, using a variable caused a short dump.

=> Added explicit conversions where the state used another type than the logger. The conversions are validated by new unit tests.

----------------------------- Demo packages ----------------------------- 
Added package /USI/BAL_DEMO (and subpackages). 
The demos can be started using the transaction codes /USI/BAL_DEMO_01 and /USI/BAL_DEMO_02.

----------------------------- Refactoring ----------------------------- 
The logging API is using the package /USI/EXCEPTION, that now offers an exception text getter. 

As this text getter did not exist when the API was written, the exception mappers in package /USI/BAL_EXCEPTION_MAPPERS were implementing the text extraction logic redundantly.

=> The redundancy was removed. The logging API will now delegate this task to the exception API.

----------------------------- Usability - Error message -----------------------------
When trying to save an empty log, the API internally raised an exception having the text "Log must be saved first!", which was rather misleading for the developer when analyzing the issue.

From now on the API will internally raise an exception having the text "The log is empty and will not be saved. Please check the log level.".

This does not change the behavior of the API - it just makes debugging a litte easier.

----------------------------- Usability - Client dependency of central cust tables ----------------------------- 
The following tables were made client dependent by "accident":
- /USI/BAL_LV_LOBJ
- /USI/BAL_LV_DATA
- /USI/BAL_LV_RTIM

The log objects in BALOBJ are client independent, just as the source code, that uses the logging API. The named tables are supposed to be maintained by the developer at design time. 

The customizing entries represent the default logging strategy defined by the developer - so they are a part of the development and should therefore be client independent.

As it made no sense for these tables to be client-dependent, this flaw has been fixed.