Issues #1, #7 / abaplint-fixes
# Bugfixes
Data containers were saved in database table /USI/BAL_DATA **WITHOUT** Compression. The amount of data in the database was thereby about twenty times as large as technically required (Issue #7).

# Enhancements
* Issue #1: The ITAB container /USI/CL_BAL_DC_ITAB can now handle internal tables having:
  * a non-DDIC line type (TYPES / DATA: BEGIN OF ...)
  * an elementary line type (String, Data element)
* Added XML-Container /USI/CL_BAL_DC_XML.

# Additional changes
## Overhaul of the de-/serialization logic of all data containers
Previously, the data was serialized as XML. 

The new version uses XML, JSON or a mixture of both to serialize a data container. 

The documents created this way are smaller than in the previous version and save - in addition to the compression (Issue #7) - further space on the database.

## Fixed various abaplint findings:
* Line breaks
* Indentation
* Removed obsolete statements (CLASS ... DEFINITION LOAD, TYPE-POOLS)
* Removed Aliases in private components of the API
* Removed unnecessary chaining
* Removed parameter name for simporting parameters where possible
* Replaced IF-ELSE to set boolean with boolc( )